fixes for CI issues
